Create Order Modules, Order Entity, and set up resolvers and services for orders

+++ src/restaurants/entities/dish.entity.ts
@@ -1,4 +1,4 @@
-import { Field, InputType, Int, ObjectType } from '@nestjs/graphql';
+import { Field, Float, InputType, ObjectType } from '@nestjs/graphql';
 import { IsNumber, IsString, Length } from 'class-validator';
 import { CoreEntity } from 'src/common/entities/core.entity';
 import { Column, Entity, ManyToOne, RelationId } from 'typeorm';
@@ -9,7 +9,7 @@ import { Restaurant } from './restaurant.entity';
 class DishChoices {
   @Field(type => String)
   name: string;
-  @Field(type => Int, { nullable: true })
+  @Field(type => Float, { nullable: true })
   extra?: number;
 }
 
@@ -22,7 +22,7 @@ class DishOption {
   @Field(type => [DishChoices], { nullable: true })
   choices?: DishChoices[];
 
-  @Field(type => Int, { nullable: true })
+  @Field(type => Float, { nullable: true })
   extra?: number;
 }
 
@@ -36,7 +36,7 @@ export class Dish extends CoreEntity {
   @Length(5)
   name: string;
 
-  @Field(type => Int)
+  @Field(type => Float)
   @Column()
   @IsNumber()
   price: number;
